Move CtInterfaceTest contents to InterfaceTest
The latter doesn't adhere to the normal conventions but already existed.
I missed it when writing my own tests.
